WIP: Working setting sync
This works, but I kind of don't like it. I wonder if there's a way to add a pending state for each of the Svelte options - that would be much better. Even if it's sparsly used, long term we might want things pending to block the UI while something is updating.