Phan: Add phan configs to all php projects (#36332)
Per pdWQjU-IH-p2, we're going to use Phan for PHP static analysis.

This commit adds Phan configs to all the php-using projects, and to new
projects created with `jetpack generate`.

It also adds a global production-exclude for the `.phan/` directory, as
we'll never want that in mirror repos.

And finally, it adds a structure test to ensure php-using projects
continue to have phan configs.

---------

Co-authored-by: Igor Zinovyev <zinigor@gmail.com>
Co-authored-by: Steve D <33553323+sdixon194@users.noreply.github.com>
Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8254939655